% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_endpoint.R
\name{get_endpoint_status}
\alias{get_endpoint_status}
\title{Get API endpoint status}
\usage{
get_endpoint_status(endpoint)
}
\arguments{
\item{endpoint}{An atomic character vector, the URL of API endpoint that needs to be checked}
}
\value{
An atomic character vector, the status code of the API endpoint
}
\description{
Get API endpoint status
}
\details{
This function is not exported
}
