% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_endpoint.R
\name{test_api_endpoint}
\alias{test_api_endpoint}
\title{Check the health of API endpoints}
\usage{
test_api_endpoint(endpoint, endpoint_name = NULL, response_time_limit = 5)
}
\arguments{
\item{endpoint}{An atomic character vector, the URL of API endpoint that needs to be checked}

\item{endpoint_name}{An atomic character vector, the name, saved as R environment variables
locally, of the API endpoint}

\item{response_time_limit}{The maximum response time for the API endpoints, beyond which the endpoint
will be considered unhealthy.}
}
\description{
Check the health of API endpoints
}
\details{
This function is exported for external use
}
